+++ packages/app/src/category/service/category.service.ts
@@ -1,4 +1,4 @@
-import { CategoryCreateEntityInterface, CategoryEntityInterface } from '@budgie/contracts';
+import { CategoryCreateEntityInterface, CategoryEntityInterface, transactionAsync } from '@budgie/contracts';
 
 import { categoryRepository, db } from '../../@generic/drizzle/db/db';
 
@@ -8,7 +8,8 @@ class CategoryService {
         for (let i = 0; i < inputs.length; i += batchSize) {
             const batch = inputs.slice(i, i + batchSize);
 
-            results.push(...(await db.transaction(async tx => categoryRepository.bulkCreate(batch, tx))));
+            // eslint-disable-next-line no-await-in-loop
+            results.push(...(await transactionAsync(db, async tx => categoryRepository.bulkCreate(batch, tx))));
         }
 
         return results.reduce<Record<string, CategoryEntityInterface>>((acc, category) => ({ ...acc, [category.title]: category }), {});