+++ packages/app/src/transaction/components/transaction-category-filter/transaction-category-filter-item.tsx
@@ -2,6 +2,7 @@ import { CategoryEntityInterface, UserIconNameEnum } from '@budgie/contracts';
 import { cva } from 'class-variance-authority';
 import { Text } from 'react-native';
 
+import { TransactionFiltersSelectors } from '../../../@e2e/selectors/transaction-filters.selector';
 import { CircleIcon } from '../../../@generic/component/circle-icon/circle-icon';
 import { HapticPressable } from '../../../@generic/component/haptic-pressable/haptic-pressable';
 import { Icon } from '../../../@generic/component/icon/icon';
@@ -35,10 +36,21 @@ export const TransactionCategoryFilterItem = ({ onSelect, category, isSelected }
     const handleSelect = () => void onSelect(category.id);
 
     return (
-        <HapticPressable onPress={handleSelect} className={categoryVariants({ isSelected })}>
+        <HapticPressable
+            onPress={handleSelect}
+            className={categoryVariants({ isSelected })}
+            testID={TransactionFiltersSelectors.CategoryOption(category.title)}
+        >
             <CircleIcon icon={category.icon} variant="ghost" size={26} iconSize={12} />
             <Text className={cn(textVariants({ isSelected }), 'mr-auto')}>{category.title}</Text>
-            {isSelected ? <Icon size={16} icon={UserIconNameEnum.Check} className="text-primary" /> : null}
+            {isSelected ? (
+                <Icon
+                    size={16}
+                    icon={UserIconNameEnum.Check}
+                    className="text-primary"
+                    testID={TransactionFiltersSelectors.CategoryOptionSelected(category.title)}
+                />
+            ) : null}
         </HapticPressable>
     );
 };