+++ packages/contracts/src/account-balance/repository/account-balance.repository.ts
@@ -1,7 +1,7 @@
 /* eslint-disable max-lines -- Repository with complex SQL aggregation queries */
 import { and, eq, inArray, isNull, sql } from 'drizzle-orm';
 
-import { DB, TX } from '../../@generic/type/db.type';
+import { DB } from '../../@generic/type/db.type';
 import { getDirectExchangeRateSql, getInverseExchangeRateSql } from '../../@generic/util/get-exchange-rate-sql.util';
 import { AccountDebtTypeEnum } from '../../account/enum/account-debt-type.enum';
 import { AccountTypeEnum } from '../../account/enum/account-type.enum';
@@ -18,7 +18,7 @@ export class AccountBalanceRepository {
     constructor(private db: DB) {}
 
     // TODO: change to bulkUpsert when drizzle is updated to the latest version
-    async upsert(input: AccountBalanceCreateEntityInterface, tx?: TX): Promise<AccountBalanceEntityInterface> {
+    async upsert(input: AccountBalanceCreateEntityInterface, tx?: DB): Promise<AccountBalanceEntityInterface> {
         const [accountBalance] = await (tx ?? this.db)
             .insert(AccountBalanceEntityTable)
             .values([input])
@@ -31,12 +31,18 @@ export class AccountBalanceRepository {
         return accountBalance;
     }
 
-    async getByAccountIds(accountIds: number[]): Promise<AccountBalanceEntityInterface[]> {
-        return await this.db.select().from(AccountBalanceEntityTable).where(inArray(AccountBalanceEntityTable.accountId, accountIds));
+    async getByAccountIds(accountIds: number[], tx?: DB): Promise<AccountBalanceEntityInterface[]> {
+        return await (tx ?? this.db).select().from(AccountBalanceEntityTable).where(inArray(AccountBalanceEntityTable.accountId, accountIds));
     }
 
-    async getNewTransactionEntriesDeltas(accountIds: number[]): Promise<Map<number, number>> {
-        const results = await this.db
+    getAllBalances() {
+        return this.db.select().from(AccountBalanceEntityTable);
+    }
+
+    async getNewTransactionEntriesDeltas(accountIds: number[], tx?: DB): Promise<Map<number, number>> {
+        const database = tx ?? this.db;
+
+        const results = await database
             .select({
                 accountId: TransactionEntryEntityTable.accountId,
                 delta: this.getTransactionsSumSql().mapWith(Number)
@@ -295,7 +301,7 @@ export class AccountBalanceRepository {
             .limit(1);
     }
 
-    async truncate(tx?: TX): Promise<void> {
+    async truncate(tx?: DB): Promise<void> {
         await (tx ?? this.db).delete(AccountBalanceEntityTable);
     }