+++ packages/contracts/src/transaction-entry/repository/transaction-entry.repository.ts
@@ -2,7 +2,7 @@ import { and, eq, inArray } from 'drizzle-orm';
 
 import { isNotEmptyArray } from '@rnw-community/shared';
 
-import { DB, TX } from '../../@generic/type/db.type';
+import { DB } from '../../@generic/type/db.type';
 import { TransactionEntryCreateEntityInterface } from '../entity/transaction-entry-create-entity.interface';
 import { TransactionEntryEntityTable } from '../table/transaction-entry-entity.table';
 
@@ -11,7 +11,7 @@ import type { TransactionEntryEntityInterface } from '../entity/transaction-entr
 export class TransactionEntryRepository {
     constructor(private db: DB) {}
 
-    async bulkCreate(inputs: TransactionEntryCreateEntityInterface[], tx?: TX): Promise<TransactionEntryEntityInterface[]> {
+    async bulkCreate(inputs: TransactionEntryCreateEntityInterface[], tx?: DB): Promise<TransactionEntryEntityInterface[]> {
         if (isNotEmptyArray(inputs)) {
             return await (tx ?? this.db).insert(TransactionEntryEntityTable).values(inputs).returning();
         }
@@ -19,13 +19,13 @@ export class TransactionEntryRepository {
         return [];
     }
 
-    async create(input: TransactionEntryCreateEntityInterface, tx?: TX): Promise<TransactionEntryEntityInterface> {
+    async create(input: TransactionEntryCreateEntityInterface, tx?: DB): Promise<TransactionEntryEntityInterface> {
         const [transactionEntry] = await (tx ?? this.db).insert(TransactionEntryEntityTable).values([input]).returning();
 
         return transactionEntry;
     }
 
-    async updateById(id: number, input: Partial<TransactionEntryCreateEntityInterface>, tx?: TX): Promise<TransactionEntryEntityInterface> {
+    async updateById(id: number, input: Partial<TransactionEntryCreateEntityInterface>, tx?: DB): Promise<TransactionEntryEntityInterface> {
         const [transactionEntry] = await (tx ?? this.db)
             .update(TransactionEntryEntityTable)
             .set(input)
@@ -39,7 +39,7 @@ export class TransactionEntryRepository {
         externalId: string,
         accountId: number,
         input: Partial<TransactionEntryCreateEntityInterface>,
-        tx?: TX
+        tx?: DB
     ): Promise<TransactionEntryEntityInterface | undefined> {
         const [transactionEntry] = await (tx ?? this.db)
             .update(TransactionEntryEntityTable)
@@ -50,11 +50,11 @@ export class TransactionEntryRepository {
         return transactionEntry;
     }
 
-    async deleteByTransactionId(transactionId: number, tx?: TX): Promise<void> {
+    async deleteByTransactionId(transactionId: number, tx?: DB): Promise<void> {
         await (tx ?? this.db).delete(TransactionEntryEntityTable).where(eq(TransactionEntryEntityTable.transactionId, transactionId));
     }
 
-    async deleteByTransactionIds(transactionIds: number[], tx?: TX): Promise<void> {
+    async deleteByTransactionIds(transactionIds: number[], tx?: DB): Promise<void> {
         if (isNotEmptyArray(transactionIds)) {
             await (tx ?? this.db)
                 .delete(TransactionEntryEntityTable)
@@ -62,25 +62,25 @@ export class TransactionEntryRepository {
         }
     }
 
-    async truncate(tx?: TX): Promise<void> {
+    async truncate(tx?: DB): Promise<void> {
         await (tx ?? this.db).delete(TransactionEntryEntityTable);
     }
 
-    async archiveByAccountIds(accountIds: number[], tx?: TX): Promise<void> {
+    async archiveByAccountIds(accountIds: number[], tx?: DB): Promise<void> {
         await (tx ?? this.db)
             .update(TransactionEntryEntityTable)
             .set({ deletedAt: new Date() })
             .where(inArray(TransactionEntryEntityTable.accountId, accountIds));
     }
 
-    async restoreByAccountIds(accountIds: number[], tx?: TX): Promise<void> {
+    async restoreByAccountIds(accountIds: number[], tx?: DB): Promise<void> {
         await (tx ?? this.db)
             .update(TransactionEntryEntityTable)
             .set({ deletedAt: null })
             .where(inArray(TransactionEntryEntityTable.accountId, accountIds));
     }
 
-    async deleteByAccountId(accountId: number, tx?: TX): Promise<void> {
+    async deleteByAccountId(accountId: number, tx?: DB): Promise<void> {
         await (tx ?? this.db).delete(TransactionEntryEntityTable).where(eq(TransactionEntryEntityTable.accountId, accountId));
     }
 }